Move NoteDatabase into its own file
